requirements.txt
--extra-index-url https://test.pypi.org/simple/
digitalai-release-sdk == 0.0.17
requests >= 2.28.1